Escaped percentage sign is not comment
As the title says, if the % sign is escaped, it is not considered as beginning of the comment section.
E.g. try this one:
bank X offers $3\%$  interest rate

This supposed to be non-changed, but the current output is:
bank X offers $3\

A negative lookbehind in the regexp could solve it, hence the proposed pull request.